chore: move around typings
